app/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langgraph>=1.0.6",
    "pinecone-client>=6.0.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
